refactor: end index component name
